Revert "Teach PackWriter how to reuse an existing object list"
This reverts commit e24f8c335782e8bef34d943f63a340995d3454b1 [formerly f5fe2dca3cb9f57891e1a4b18832fcc158d0c490].

I regret adding this feature to the public API.  Caches aren't always
the best idea, as they require work to maintain.  Here the cache is
redundant information that must be computed, and when it grows stale
must be removed.  The redundant information takes up more disk space,
about the same size as the pack-*.idx files are.  For the linux-2.6
repository, that's more than 40 MB for a 400 MB repository.  So the
cache is a 10% increase in disk usage.

The entire point of this cache is to improve PackWriter performance,
and only PackWriter performance, and only when sending an initial
clone to a new client.  There may be better ways to optimize this, and
until we have a solid solution, we shouldn't be using a separate cache
in JGit.


Former-commit-id: 065a0a8122be356d1f898321763d3518b504b075